Listos los CRUDs de Cliente, Administrador y Producto con el tratamiento de imagenes

+++ app/Http/Controllers/ProductoController.php
@@ -4,27 +4,26 @@ namespace App\Http\Controllers;
 
 use App\Models\Producto;
 use Illuminate\Http\Request;
-
-use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use MongoDB\BSON\ObjectId;
 
 class ProductoController extends Controller
 {
-    use AuthorizesRequests;
+    // use AuthorizesRequests;
 
     public function index()
     {
-      try {
-          $productos = Producto::all();
-          return response()->json([
+        try {
+            $productos = Producto::all();
+            return response()->json([
             'message' => 'Productos recuperados con exito.',
             'productos'=> $productos
         ], 200);
-      } catch (\Exception $e){
+        } catch (\Exception $e){
         return response()->json([
             'message' => 'Error al recuperar los productos.',
             'productos'=> $e->getMessage()
         ], 200);
-      }
+        }
     }
 
     public function store(Request $request)
@@ -34,18 +33,18 @@ class ProductoController extends Controller
             $validated = $request->validate([
                 'nombre' => 'required|string|max:50',
                 'descripcion' => 'required|string|max:255',
-                'categoria' => 'required', // Almacenar la categoría como un arreglo dentro del documento
+                'categoria' => 'required',
                 'precio' => 'required|numeric|min:0',
                 'estado' => 'required|string|in:Activo,Inactivo,Existente,No Existente',
                 'existencia' => 'required|integer|min:0',
-                'imagen_1' => 'nullable|max:2048',
-                'imagen_2' => 'nullable|max:2048',
-                'imagen_3' => 'nullable|max:2048',
+                'imagen1' => 'nullable|image|max:2048',
+                'imagen2' => 'nullable|image|max:2048',
+                'imagen3' => 'nullable|image|max:2048',
             ]);
 
             $producto = new Producto($validated);
 
-            foreach (['imagen_1', 'imagen_2', 'imagen_3'] as $key) {
+            foreach (['imagen1', 'imagen2', 'imagen3'] as $key) {
                 if ($request->hasFile($key)) {
                     $img = $request->file($key);
                     $path = $img->storeAs('imagenes/productos', "{$key}_" . time() . ".{$img->extension()}", 'public');
@@ -67,7 +66,6 @@ class ProductoController extends Controller
     public function show($id)
     {
         $producto = Producto::findOrFail($id);
-        // $this->authorize('view', $producto);
 
         return response()->json($producto, 200);
     }
@@ -75,7 +73,6 @@ class ProductoController extends Controller
     public function update(Request $request, $id)
     {
         $producto = Producto::findOrFail($id);
-        // $this->authorize('update', $producto);
 
         try {
             $validated = $request->validate([
@@ -85,14 +82,14 @@ class ProductoController extends Controller
                 'precio' => 'required|numeric|min:0',
                 'estado' => 'required|string|in:Activo,Inactivo,Existente,No Existente',
                 'existencia' => 'required|integer|min:0',
-                'imagen_1' => 'nullable|max:2048',
-                'imagen_2' => 'nullable|max:2048',
-                'imagen_3' => 'nullable|max:2048',
+                'imagen1' => 'nullable|max:2048',
+                'imagen2' => 'nullable|max:2048',
+                'imagen3' => 'nullable|max:2048',
             ]);
 
             $producto->update($validated);
 
-            foreach (['imagen_1', 'imagen_2', 'imagen_3'] as $key) {
+            foreach (['imagen1', 'imagen2', 'imagen3'] as $key) {
                 if ($request->hasFile($key)) {
                     $img = $request->file($key);
                     $path = $img->storeAs('imagenes/productos', "{$key}_" . time() . ".{$img->extension()}", 'public');
@@ -110,10 +107,33 @@ class ProductoController extends Controller
 
     public function destroy($id)
     {
-        $producto = Producto::findOrFail($id);
-        // $this->authorize('delete', $producto);
-        $producto->delete();
+        try {
+            $objectId = new ObjectId($id);
 
-        return response()->json(['message' => 'Producto eliminado con éxito.'], 204);
+            $producto = Producto::where('_id', $objectId)->first();
+
+            if (!$producto) {
+                return response()->json([
+                    'message' => 'Producto no encontrado',
+                    'productos' => null
+                ], 404);
+            }
+            $producto->delete();
+
+            return response()->json(
+                [
+                    'message' => 'producto eliminado',
+                    'productos' => 'eliminado'
+                ],
+                200
+            );
+
+
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => 'error al eliminar producto',
+                'productos' => $e->getMessage()
+            ], 500);
+        }
     }
 }
